internal package cleanup - moved stuff around to make more sense
