+++ internal/command/playlist_example.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/gotracker/gotracker/internal/command/internal/playlist"
+	"github.com/gotracker/gotracker/internal/playlist"
 	"github.com/heucuva/optional"
 
 	"github.com/spf13/cobra"
