+++ internal/profiling/allowed.go
